+++ tests/LayerPhaseTest/src/LayerPhaseTestProbe.cpp
@@ -48,7 +48,7 @@ int LayerPhaseTestProbe::outputState(double timed) {
    }
    for (int b = 0; b < parent->getNBatch(); b++) {
       if (timed >= equilibriumTime) {
-         double tol = 1e-6;
+         float tol = 1e-6;
          FatalIf(!(fabs(fMin[b] - equilibriumValue) < tol), "Test failed.\n");
          FatalIf(!(fabs(fMax[b] - equilibriumValue) < tol), "Test failed.\n");
          FatalIf(!(fabs(avg[b] - equilibriumValue) < tol), "Test failed.\n");
